+++ dotnet/Core/Database/Server/Core/Hosting/AllorsServerApplicationBuilderExtensions.cs
@@ -1,0 +1,79 @@
+// <copyright file="AllorsServerApplicationBuilderExtensions.cs" company="Allors bv">
+// Copyright (c) Allors bv. All rights reserved.
+// Licensed under the LGPL license. See LICENSE file in the project root for full license information.
+// </copyright>
+
+namespace Allors.Server
+{
+    using Allors.Services;
+    using Database.Adapters;
+    using Database.Configuration;
+    using Database.Configuration.Derivations.Default;
+    using Database.Domain;
+    using Database.Meta;
+    using JSNLog;
+    using Microsoft.AspNetCore.Builder;
+    using Microsoft.AspNetCore.Hosting;
+    using Microsoft.Extensions.Configuration;
+    using Microsoft.Extensions.DependencyInjection;
+    using Microsoft.Extensions.Hosting;
+    using Microsoft.Extensions.Logging;
+    using ObjectFactory = Database.ObjectFactory;
+    using User = Database.Domain.User;
+
+    public static class AllorsServerApplicationBuilderExtensions
+    {
+        public static void UseAllorsServer(this IApplicationBuilder app, IWebHostEnvironment env, ILoggerFactory loggerFactory)
+        {
+            // Allors
+            var metaPopulation = new MetaBuilder().Build();
+            var engine = new Engine(Rules.Create(metaPopulation));
+            var objectFactory = new ObjectFactory(metaPopulation, typeof(User));
+            var databaseScope = new DefaultDatabaseServices(engine);
+            var configuration = app.ApplicationServices.GetRequiredService<IConfiguration>();
+            var databaseBuilder = new DatabaseBuilder(databaseScope, configuration, objectFactory, null, 60);
+            var databaseService = app.ApplicationServices.GetRequiredService<IDatabaseService>();
+            databaseService.Build = () => databaseBuilder.Build();
+            databaseService.Database = databaseService.Build();
+
+            if (env.IsDevelopment())
+            {
+                app.UseDeveloperExceptionPage();
+            }
+            else
+            {
+                app.UseHsts();
+            }
+
+            app.UseCors();
+
+            var jsnlogConfiguration = new JsnlogConfiguration
+            {
+                corsAllowedOriginsRegex = ".*",
+                serverSideMessageFormat = env.IsDevelopment() ?
+                                            "%requestId | %url | %message" :
+                                            "%requestId | %url | %userHostAddress | %userAgent | %message",
+            };
+
+            app.UseJSNLog(new LoggingAdapter(loggerFactory), jsnlogConfiguration);
+
+            // app.UseHttpsRedirection();
+            app.UseRouting();
+            app.UseAuthentication();
+            app.UseAuthorization();
+
+            app.ConfigureExceptionHandler(env);
+            app.UseResponseCaching();
+
+            app.UseMiddleware<ClaimsPrincipalServiceMiddleware>();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllerRoute(
+                    name: "default",
+                    pattern: "allors/{controller=Home}/{action=Index}/{id?}");
+                endpoints.MapControllers();
+            });
+        }
+    }
+}